+++ src/main/java/roomescape/controller/TimeController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import roomescape.domain.Time;
-import roomescape.dto.TimeRequest;
-import roomescape.dto.TimeResponse;
+import roomescape.model.domain.Time;
+import roomescape.model.dto.request.TimeRequest;
+import roomescape.model.dto.response.TimeResponse;
 import roomescape.service.TimeService;
 
 @Controller
