+++ src/main/java/roomescape/model/domain/Time.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.model.domain;
 
 import java.time.LocalTime;
 import lombok.EqualsAndHashCode;
