+++ src/main/java/roomescape/model/dto/request/TimeRequest.java
@@ -1,9 +1,9 @@
-package roomescape.dto;
+package roomescape.model.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
-import roomescape.domain.Time;
+import roomescape.model.domain.Time;
 
 public record TimeRequest(
         @NotNull(message = "시간은 필수입니다.")
