+++ src/test/java/roomescape/dao/TimeDAOTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.domain.Time;
+import roomescape.model.domain.Time;
 import roomescape.exception.NotFoundException;
 
 @JdbcTest
